Rename fields related to cmd stdout
